requirements.txt
beautifulsoup4>=4.8,<4.9
requests>=2.25,<2.26
youtube_dl==2021.4.7
django-environ>=0.4,<0.5
